StartNewMessageButtonRemoved
